Allow reconnection to lycan server. Provide errors infos to client.
